requirements.txt
amqp==5.2.0
annotated-types==0.7.0
anyio==4.6.2.post1
APScheduler==3.10.4
asgiref==3.8.1
async-timeout==4.0.3
billiard==4.2.1
cachetools==5.5.0
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
drf_yasg
Django==5.1.2
django-apscheduler==0.7.0
django-rest-swagger==2.2.0
djangorestframework==3.15.2
exceptiongroup==1.2.2
fastapi==0.110.1
google-api-core==2.21.0
google-api-python-client==2.125.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.65.0
greenlet==3.1.1
h11==0.14.0
httplib2==0.22.0
idna==3.10
itypes==1.2.0
Jinja2==3.1.4
kombu==5.4.2
MarkupSafe==3.0.2
openapi-codec==1.3.2
prompt_toolkit==3.0.48
proto-plus==1.24.0
protobuf==5.28.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
redis==5.0.3
requests==2.31.0
rsa==4.9
simplejson==3.19.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
sqlparse==0.5.1
starlette==0.37.2
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.29.0
vine==5.1.0
wcwidth==0.2.13
